Prepare for C version
